Moved sidebar to own component
