+++ src/components/UserresearchSidebar.tsx
@@ -1,0 +1,78 @@
+import React from 'react';
+import { Sidebar } from '@tpr/layout';
+import { useLocation, useHistory, matchPath } from 'react-router-dom';
+
+const UserresearchSidebar = () => {
+  const history = useHistory();
+  const location = useLocation();
+  return (
+    <Sidebar
+      history={history}
+      location={location}
+      title={'Scheme return home'}
+      matchPath={matchPath}
+      sections={[
+        {
+          title: 'Scheme details',
+          order: 1,
+          links: [
+            {
+              name: 'Scheme name and address',
+              completed: true,
+              path: '/scheme-name',
+            },
+            {
+              name: 'Scheme status and membership',
+              completed: false,
+              path: '/scheme-status-and-membership',
+            },
+            {
+              name: 'Consent to electronic communication',
+              completed: false,
+              path: '/electronic-communication',
+            },
+          ],
+        },
+        {
+          title: 'Role',
+          order: 2,
+          links: [
+            {
+              name: 'Trustee details',
+              completed: false,
+              path: '/trustee-details',
+            },
+            {
+              name: 'Employer details',
+              completed: false,
+              path: '/employer-details',
+            },
+            {
+              name: 'Service provider details',
+              completed: false,
+              path: '/service-provider-details',
+            },
+            {
+              name: 'Named contact details',
+              completed: false,
+              path: '/named-contract-details',
+            },
+          ],
+        },
+        {
+          title: 'Finish Up',
+          order: 3,
+          links: [
+            {
+              name: 'Review and submit',
+              completed: false,
+              path: '/review-and-submit',
+            },
+          ],
+        },
+      ]}
+    />
+  );
+};
+
+export default UserresearchSidebar;
