+++ src/views/SchemeStatusAndMembership.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { H1, Hr, P, Flex, H4 } from '@tpr/core';
 import Styles from './Layout.module.scss';
-import { Sidebar, ArrowLink, ArrowButton } from '@tpr/layout';
+import { ArrowLink, ArrowButton } from '@tpr/layout';
 import {
   Form,
   renderFields,
@@ -10,12 +10,9 @@ import {
   SeparatorY,
   validate,
 } from '@tpr/forms';
-import { useLocation, useHistory, matchPath } from 'react-router-dom';
+import UserresearchSidebar from '../components/UserresearchSidebar';
 
 const SchemeStatusAndMembership = () => {
-  const history = useHistory();
-  const location = useLocation();
-
   const RADIO_BUTTON_NAME = 'schemeStatus';
 
   const SchemeStatusFields: FieldProps[] = [
@@ -129,72 +126,7 @@ const SchemeStatusAndMembership = () => {
 
   return (
     <div className={Styles.root}>
-      <Sidebar
-        history={history}
-        location={location}
-        title={'Scheme return home'}
-        matchPath={matchPath}
-        sections={[
-          {
-            title: 'Scheme details',
-            order: 1,
-            links: [
-              {
-                name: 'Scheme name and address',
-                completed: true,
-                path: '/scheme-name',
-              },
-              {
-                name: 'Scheme status and membership',
-                completed: false,
-                path: '/scheme-status-and-membership',
-              },
-              {
-                name: 'Consent to electronic communication',
-                completed: false,
-                path: '/electronic-communication',
-              },
-            ],
-          },
-          {
-            title: 'Role',
-            order: 2,
-            links: [
-              {
-                name: 'Trustee details',
-                completed: false,
-                path: '/trustee-details',
-              },
-              {
-                name: 'Employer details',
-                completed: false,
-                path: '/employer-details',
-              },
-              {
-                name: 'Service provider details',
-                completed: false,
-                path: '/service-provider-details',
-              },
-              {
-                name: 'Named contact details',
-                completed: false,
-                path: '/named-contract-details',
-              },
-            ],
-          },
-          {
-            title: 'Finish Up',
-            order: 3,
-            links: [
-              {
-                name: 'Review and submit',
-                completed: false,
-                path: '/review-and-submit',
-              },
-            ],
-          },
-        ]}
-      />
+      <UserresearchSidebar />
 
       <div className={Styles.main}>
         <ArrowLink
